% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyhydat.ws-package.R
\docType{package}
\name{tidyhydat.ws-package}
\alias{tidyhydat.ws}
\alias{tidyhydat.ws-package}
\title{tidyhydat.ws: Access ECCC Webservice}
\description{
A stand alone package that is a client to the ECCC webservice.
}
\author{
\strong{Maintainer}: Sam Albers \email{sam.albers@gov.bc.ca}

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
